Change client-side API to support PBs as params by default.
Since we still have legacy clients, we need to keep the old API around
and test it for a while, hence the tests of the old APIs in here.

+++ src/main/java/org/retrostore/RetrostoreClient.java
@@ -66,8 +66,8 @@ public interface RetrostoreClient {
    *
    * @return A unique token that can be used to fetch this state later.
    */
-  long uploadState(SystemState state) throws ApiException;
+  int uploadState(SystemState state) throws ApiException;
 
   /** Fetches a system state associated with the given token. */
-  SystemState downloadState(long token) throws ApiException;
+  SystemState downloadState(int token) throws ApiException;
 }
